pyproject.toml
[tool.poetry]
name = "coin-trader"
version = "0.1.0"
description = "안전성 최우선 코인 자동거래 시스템"
authors = ["seunghwan"]
packages = [{include = "coin_trader"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
httpx = "^0.27.0"
tenacity = "^8.2.3"
cryptography = "^42.0.0"
structlog = "^24.1.0"
python-dotenv = "^1.0.0"
numpy = "^1.26.0"
PyJWT = "^2.8.0"
click = "^8.1.0"
fastapi = "^0.115.0"
uvicorn = {version = "^0.34.0", extras = ["standard"]}
jinja2 = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.2.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
trader = "coin_trader.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
